pyproject.toml
[project]
name = "nasa-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.2",
    "aiosignal==1.4.0",
    "altair==5.5.0",
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "appdirs==1.4.4",
    "argon2-cffi==25.1.0",
    "argon2-cffi-bindings==25.1.0",
    "arrow==1.3.0",
    "asttokens==3.0.0",
    "async-lru==2.0.5",
    "attrs==25.3.0",
    "babel==2.17.0",
    "backoff==2.2.1",
    "bcrypt==4.3.0",
    "beautifulsoup4==4.13.4",
    "bleach==6.2.0",
    "blinker==1.9.0",
    "build==1.3.0",
    "cachetools==5.5.2",
    "certifi==2025.8.3",
    "cffi==1.17.1",
    "cfgv==3.5.0",
    "charset-normalizer==3.4.3",
    "chromadb==1.0.16",
    "click==8.2.1",
    "coloredlogs==15.0.1",
    "comm==0.2.3",
    "dataclasses-json==0.6.7",
    "datasets==4.4.2",
    "debugpy==1.8.16",
    "decorator==5.2.1",
    "defusedxml==0.7.1",
    "dill==0.4.0",
    "diskcache==5.6.3",
    "distlib==0.4.0",
    "distro==1.9.0",
    "docstring-parser==0.17.0",
    "durationpy==0.10",
    "executing==2.2.0",
    "fastjsonschema==2.21.1",
    "filelock==3.18.0",
    "flatbuffers==25.2.10",
    "fqdn==1.5.1",
    "frozenlist==1.8.0",
    "fsspec==2025.7.0",
    "gitdb==4.0.12",
    "gitpython==3.1.45",
    "google-auth==2.40.3",
    "googleapis-common-protos==1.70.0",
    "greenlet==3.3.0",
    "grpcio==1.74.0",
    "h11==0.16.0",
    "hf-xet==1.1.7",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "huggingface-hub==0.34.4",
    "humanfriendly==10.0",
    "identify==2.6.15",
    "idna==3.10",
    "importlib-metadata==8.7.0",
    "importlib-resources==6.5.2",
    "instructor==1.13.0",
    "ipykernel==6.29.5",
    "ipython==9.4.0",
    "ipython-pygments-lexers==1.1.1",
    "ipywidgets==8.1.7",
    "isoduration==20.11.0",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "jiter==0.10.0",
    "json5==0.12.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.25.0",
    "jsonschema-specifications==2025.4.1",
    "jupyter==1.1.1",
    "jupyter-client==8.6.3",
    "jupyter-console==6.6.3",
    "jupyter-core==5.8.1",
    "jupyter-events==0.12.0",
    "jupyter-lsp==2.2.6",
    "jupyter-server==2.16.0",
    "jupyter-server-terminals==0.5.3",
    "jupyterlab==4.4.5",
    "jupyterlab-pygments==0.3.0",
    "jupyterlab-server==2.27.3",
    "jupyterlab-widgets==3.0.15",
    "kubernetes==33.1.0",
    "langchain==1.2.0",
    "langchain-classic==1.0.1",
    "langchain-community==0.4.1",
    "langchain-core==1.2.5",
    "langchain-openai==1.1.6",
    "langchain-text-splitters==1.1.0",
    "langgraph==1.0.5",
    "langgraph-checkpoint==3.0.1",
    "langgraph-prebuilt==1.0.5",
    "langgraph-sdk==0.3.1",
    "langsmith==0.5.2",
    "lark==1.2.2",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.2",
    "marshmallow==3.26.2",
    "matplotlib-inline==0.1.7",
    "mdurl==0.1.2",
    "mistune==3.1.3",
    "mmh3==5.2.0",
    "mpmath==1.3.0",
    "multidict==6.7.0",
    "multiprocess==0.70.18",
    "mypy-extensions==1.1.0",
    "narwhals==2.1.1",
    "nbclient==0.10.2",
    "nbconvert==7.16.6",
    "nbformat==5.10.4",
    "nest-asyncio==1.6.0",
    "networkx==3.6.1",
    "nodeenv==1.10.0",
    "notebook==7.4.5",
    "notebook-shim==0.2.4",
    "numpy==2.3.2",
    "oauthlib==3.3.1",
    "onnxruntime==1.22.1",
    "openai==2.14.0",
    "opentelemetry-api==1.36.0",
    "opentelemetry-exporter-otlp-proto-common==1.36.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.36.0",
    "opentelemetry-proto==1.36.0",
    "opentelemetry-sdk==1.36.0",
    "opentelemetry-semantic-conventions==0.57b0",
    "orjson==3.11.1",
    "ormsgpack==1.12.1",
    "overrides==7.7.0",
    "packaging==25.0",
    "pandas==2.3.0",
    "pandocfilters==1.5.1",
    "parso==0.8.4",
    "pexpect==4.9.0",
    "pillow==11.3.0",
    "platformdirs==4.3.8",
    "posthog==5.4.0",
    "pre-commit==4.5.1",
    "prometheus-client==0.22.1",
    "prompt-toolkit==3.0.51",
    "propcache==0.4.1",
    "protobuf==6.31.1",
    "psutil==7.0.0",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pyarrow==21.0.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pybase64==1.4.2",
    "pycparser==2.22",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.12.0",
    "pydeck==0.9.1",
    "pygments==2.19.2",
    "pypika==0.48.9",
    "pyproject-hooks==1.2.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "python-json-logger==3.3.0",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "pyzmq==27.0.1",
    "ragas==0.4.2",
    "rapidfuzz>=3.14.3",
    "referencing==0.36.2",
    "regex==2025.11.3",
    "requests==2.32.5",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "rfc3339-validator==0.1.4",
    "rfc3986-validator==0.1.1",
    "rfc3987-syntax==1.1.0",
    "rich==14.1.0",
    "rouge-score>=0.1.2",
    "rpds-py==0.27.0",
    "rsa==4.9.1",
    "sacrebleu>=2.5.1",
    "scikit-network==0.33.5",
    "scipy==1.16.3",
    "send2trash==1.8.3",
    "setuptools==80.9.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "soupsieve==2.7",
    "sqlalchemy==2.0.45",
    "stack-data==0.6.3",
    "streamlit==1.47.0",
    "sympy==1.14.0",
    "tenacity==9.1.2",
    "terminado==0.18.1",
    "tiktoken==0.12.0",
    "tinycss2==1.4.0",
    "tokenizers==0.21.4",
    "toml==0.10.2",
    "tornado==6.5.2",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "ty==0.0.8",
    "typer==0.16.0",
    "types-python-dateutil==2.9.0.20250809",
    "typing-extensions==4.14.1",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "uri-template==1.3.0",
    "urllib3==2.5.0",
    "uuid-utils==0.12.0",
    "uvicorn==0.35.0",
    "virtualenv==20.35.4",
    "watchdog==6.0.0",
    "watchfiles==1.1.0",
    "wcwidth==0.2.13",
    "webcolors==24.11.1",
    "webencodings==0.5.1",
    "websocket-client==1.8.0",
    "websockets==15.0.1",
    "widgetsnbextension==4.0.14",
    "xxhash==3.6.0",
    "yarl==1.22.0",
    "zipp==3.23.0",
    "zstandard==0.25.0",
]
